feat(errors): Enhance PastepwnError class and testing suite
- Add test case for handling empty messages
- Introduce test case for None messages
- Validate class's handling of custom object messages
- Strengthen Unicode message handling through testing
- Expand testing to cover integer messages and multiple arguments
